refactor(agents): remove dynamicAgents surface and module-level run(); single base44.agents namespace
