+++ src/client.ts
@@ -10,7 +10,6 @@ import {
 import { getAccessToken } from "./utils/auth-utils.js";
 import { createFunctionsModule } from "./modules/functions.js";
 import { createAgentsModule } from "./modules/agents.js";
-import { createDynamicAgentsModule } from "./modules/dynamic-agents.js";
 import { createAppLogsModule } from "./modules/app-logs.js";
 import { createUsersModule } from "./modules/users.js";
 import { RoomsSocket, RoomsSocketConfig } from "./utils/socket-utils.js";
@@ -192,10 +191,6 @@ export function createClient(config: CreateClientConfig): Base44Client {
       token,
       getToken: () => token || getAccessToken() || undefined,
     }),
-    dynamicAgents: createDynamicAgentsModule({
-      serverUrl,
-      getToken: () => token || getAccessToken() || undefined,
-    }),
     appLogs: createAppLogsModule(axiosClient, appId),
     users: createUsersModule(axiosClient, appId),
     analytics: createAnalyticsModule({
@@ -238,10 +233,6 @@ export function createClient(config: CreateClientConfig): Base44Client {
       appId,
       serverUrl,
       token,
-      getToken: () => serviceToken,
-    }),
-    dynamicAgents: createDynamicAgentsModule({
-      serverUrl,
       getToken: () => serviceToken,
     }),
     appLogs: createAppLogsModule(serviceRoleAxiosClient, appId),
